+++ lib/src/xxh3.dart
@@ -25,6 +25,9 @@ const XXH3 xxh3_64 = xxh3;
 ///
 /// [wiki]: https://github.com/Cyan4973/xxHash/wiki/Collision-ratio-comparison
 ///
+/// **Note: XXH3 is not supported on the web platform. Any usage there will
+/// throw an [UnimplementedError].**
+///
 /// **WARNING: It should not be used for cryptographic purposes.**
 class XXH3 extends BlockHashBase {
   final int seed;